Исправление замечаний CodeFactor
