+++ Zilon.Client/Assets/Zilon/Sprites/stairs.png.meta
@@ -1,11 +1,15 @@
 fileFormatVersion: 2
 guid: 3865c6a5697725f4eb02f84c6615f39e
 TextureImporter:
-  fileIDToRecycleName:
-    21300000: stairs_0
-    21300002: stairs_1
+  internalIDToNameTable:
+  - first:
+      213: 21300000
+    second: stairs_0
+  - first:
+      213: 21300002
+    second: stairs_1
   externalObjects: {}
-  serializedVersion: 7
+  serializedVersion: 10
   mipmaps:
     mipMapMode: 0
     enableMipMap: 0
@@ -60,7 +64,7 @@ TextureImporter:
   compressionQualitySet: 0
   textureFormatSet: 0
   platformSettings:
-  - serializedVersion: 2
+  - serializedVersion: 3
     buildTarget: DefaultTexturePlatform
     maxTextureSize: 2048
     resizeAlgorithm: 0
@@ -71,7 +75,8 @@ TextureImporter:
     allowsAlphaSplitting: 0
     overridden: 0
     androidETC2FallbackOverride: 0
-  - serializedVersion: 2
+    forceMaximumCompressionQuality_BC6H_BC7: 1
+  - serializedVersion: 3
     buildTarget: Standalone
     maxTextureSize: 2048
     resizeAlgorithm: 0
@@ -82,6 +87,7 @@ TextureImporter:
     allowsAlphaSplitting: 0
     overridden: 0
     androidETC2FallbackOverride: 0
+    forceMaximumCompressionQuality_BC6H_BC7: 1
   spriteSheet:
     serializedVersion: 2
     sprites:
@@ -101,6 +107,7 @@ TextureImporter:
       tessellationDetail: 0
       bones: []
       spriteID: 266ac9516ee2ddd4aa57959d8831dabd
+      internalID: 21300000
       vertices: []
       indices: 
       edges: []
@@ -121,6 +128,7 @@ TextureImporter:
       tessellationDetail: 0
       bones: []
       spriteID: 90b5e798758e6b84892bc3c2f49a93d4
+      internalID: 21300002
       vertices: []
       indices: 
       edges: []
@@ -129,10 +137,12 @@ TextureImporter:
     physicsShape: []
     bones: []
     spriteID: 98fdac3f0b41fae4b9285c4c296bb19e
+    internalID: 0
     vertices: []
     indices: 
     edges: []
     weights: []
+    secondaryTextures: []
   spritePackingTag: 
   pSDRemoveMatte: 0
   pSDShowRemoveMatteOption: 0
