+++ Zilon.Core/Zilon.Core/Schemes/IPersonSurvivalStatSubScheme.cs
@@ -12,7 +12,6 @@
         /// </summary>
         PersonSurvivalStatType Type { get; }
 
-
         /// <summary>
         /// Ключевые точки характеристики.
         /// При прохождении ключевых точек будет выбрасываться событие.
